Added: Created a dedicated transactions folder under utils/api to organize transaction-related API functions
