+++ src/components/betting-summary/BettorsvsBetsPlacedChart.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Box, Typography, Stack } from "@mui/material";
 import { BarChart } from "@mui/x-charts/BarChart";
 
-import fetchHistoricalSummary from "~/utils/api/getHistoricalSummary";
+import fetchHistoricalSummary from "~/utils/api/transactions/getHistoricalSummary";
 
 const CustomLegend = () => (
   <Stack
